[Radoub] tech-debt: Split FileBrowserPanelBase.axaml.cs into partials (#2426)
Break the 1199-line code-behind into focused partial files by existing region
boundaries: .Loading (load/filter/list pipeline), .UI (status helpers + search/
selection/sort/collapse handlers), .FileOps (delete/copy/rename + entry lookup),
.Indexing (sort columns + background metadata indexing), .CopyToModule (archive
copy flow). Main file keeps shared state, properties, ctor, IFileBrowserPanel.
Pure mechanical split, no behavior change.
Tests: Radoub.UI.Tests FileBrowser/BrowserSort/BrowserSave 55/55; full solution builds 0 errors.
